DEX-951 Multiple response layers in Test API
* DexApi moved to it.api.dex
* Introduced AsyncEnrichedApi to remove duplication
