+++ dex-it-common/src/main/scala/com/wavesplatform/dex/it/api/dex/DexApiWaitOps.scala
@@ -1,4 +1,4 @@
-package com.wavesplatform.dex.it.dex
+package com.wavesplatform.dex.it.api.dex
 
 import cats.Functor
 import cats.syntax.functor._
