+++ dex-it-common/src/main/scala/com/wavesplatform/dex/it/api/node/AsyncEnrichedNodeApi.scala
@@ -1,23 +1,21 @@
 package com.wavesplatform.dex.it.api.node
 
 import java.net.InetSocketAddress
-import java.util.UUID
 
 import com.softwaremill.sttp._
 import com.softwaremill.sttp.playJson._
 import com.wavesplatform.dex.domain.account.Address
 import com.wavesplatform.dex.domain.asset.Asset
-import com.wavesplatform.dex.it.api.node.AsyncEnrichedNodeApi._
 import com.wavesplatform.dex.it.api.responses.node._
-import com.wavesplatform.dex.it.api.{AsyncEnriched, EnrichedResponse}
+import com.wavesplatform.dex.it.api.{AsyncEnriched, AsyncEnrichedApi}
 import im.mak.waves.transactions.Transaction
 import im.mak.waves.transactions.common.Id
-import play.api.libs.json.{JsSuccess, Reads}
 
 import scala.concurrent.{ExecutionContext, Future}
 
 class AsyncEnrichedNodeApi(apiKey: String, host: => InetSocketAddress)(implicit ec: ExecutionContext, httpBackend: SttpBackend[Future, Nothing])
-    extends NodeApi[AsyncEnriched] {
+    extends AsyncEnrichedApi(host)
+    with NodeApi[AsyncEnriched] {
 
   override def wavesBalanceOrig(address: Address): AsyncEnriched[WavesBalanceResponse] = mk {
     sttp.get(uri"$apiUri/addresses/balance/$address")
@@ -52,31 +50,6 @@ class AsyncEnrichedNodeApi(apiKey: String, host: => InetSocketAddress)(implicit 
 
   override def connectedPeers: AsyncEnriched[ConnectedPeersResponse] = mk {
     sttp.get(uri"$apiUri/peers/connected")
-  }
-
-  def apiUri: String = {
-    val savedHost = host
-    s"http://${savedHost.getAddress.getHostAddress}:${savedHost.getPort}"
-  }
-
-  def mk[T: Reads](req: Request[String, Nothing]): AsyncEnriched[T] =
-    httpBackend.send[String](req.tag("requestId", UUID.randomUUID)).map {
-      EnrichedResponse.AsJson[T](_)
-    }
-
-  def mkHocon[T](req: Request[String, Nothing]): AsyncEnriched[T] =
-    httpBackend.send[String](req.tag("requestId", UUID.randomUUID)).map(EnrichedResponse.AsHocon[T])
-
-  def mkIgnore(req: Request[String, Nothing]): AsyncEnriched[Unit] =
-    httpBackend.send[String](req.tag("requestId", UUID.randomUUID)).map(EnrichedResponse.Ignore)
-
-}
-
-// TODO
-object AsyncEnrichedNodeApi {
-
-  implicit val transactionReads: Reads[Transaction] = Reads { js =>
-    JsSuccess(Transaction.fromJson(js.toString()))
   }
 
 }
