+++ dex-it/src/test/scala/com/wavesplatform/it/api/MatcherCommand.scala
@@ -2,10 +2,7 @@ package com.wavesplatform.it.api
 
 import com.wavesplatform.dex.domain.account.KeyPair
 import com.wavesplatform.dex.domain.order.Order
-import com.wavesplatform.dex.it.dex.DexApi
 import com.wavesplatform.dex.it.docker.DexContainer
-
-import scala.concurrent.Future
 
 sealed trait MatcherCommand extends Product with Serializable
